refactor(api): align v1 namespace and move conversion tests #1959
Part A: Move misplaced tests
- Move conversion unit tests from axum-rest-api-server to
  rest-api-runtime-adapter::v1::conversion (tests belong with
  the production code they exercise)
- Add torrust-clock dev-dependency to rest-api-runtime-adapter
- Remove empty peer.rs stub (only doc comment, no code)
- Clean up resources/mod.rs module declaration

Part B: Align v1 namespace
- Add v1/ module to rest-api-application, move ports/ and
  use_cases/ under it
- Add v1/ module to rest-api-runtime-adapter, move adapters/,
  container.rs, conversion.rs under it
- Update all internal crate:: paths
- Update all cross-package imports across workspace
  (axum-rest-api-server, src/bootstrap, src/container)
