+++ packages/axum-rest-api-server/src/v1/context/torrent/resources/mod.rs
@@ -1,4 +1,3 @@
 //! API resources for the [`torrent`](crate::v1::context::torrent)
 //! API context.
-pub mod peer;
 pub mod torrent;
