--- packages/axum-rest-api-server/src/v1/context/torrent/resources/peer.rs
@@ -1,3 +1,0 @@
-//! `Peer` and Peer `Id` API resources.
-//!
-//! Protocol DTOs are defined in `torrust-tracker-rest-api-protocol`.
