+++ packages/axum-rest-api-server/src/v1/routes.rs
@@ -2,15 +2,15 @@
 use std::sync::Arc;
 
 use axum::Router;
-use torrust_tracker_rest_api_application::use_cases::auth_key::AuthKeyApiService;
-use torrust_tracker_rest_api_application::use_cases::stats::StatsApiService;
-use torrust_tracker_rest_api_application::use_cases::torrent::TorrentApiService;
-use torrust_tracker_rest_api_application::use_cases::whitelist::WhitelistApiService;
-use torrust_tracker_rest_api_runtime_adapter::adapters::auth_key::TrackerAuthKeyAdapter;
-use torrust_tracker_rest_api_runtime_adapter::adapters::stats::TrackerStatsAdapter;
-use torrust_tracker_rest_api_runtime_adapter::adapters::torrent::TrackerTorrentQueryAdapter;
-use torrust_tracker_rest_api_runtime_adapter::adapters::whitelist::TrackerWhitelistAdapter;
-use torrust_tracker_rest_api_runtime_adapter::container::TrackerHttpApiCoreContainer;
+use torrust_tracker_rest_api_application::v1::use_cases::auth_key::AuthKeyApiService;
+use torrust_tracker_rest_api_application::v1::use_cases::stats::StatsApiService;
+use torrust_tracker_rest_api_application::v1::use_cases::torrent::TorrentApiService;
+use torrust_tracker_rest_api_application::v1::use_cases::whitelist::WhitelistApiService;
+use torrust_tracker_rest_api_runtime_adapter::v1::adapters::auth_key::TrackerAuthKeyAdapter;
+use torrust_tracker_rest_api_runtime_adapter::v1::adapters::stats::TrackerStatsAdapter;
+use torrust_tracker_rest_api_runtime_adapter::v1::adapters::torrent::TrackerTorrentQueryAdapter;
+use torrust_tracker_rest_api_runtime_adapter::v1::adapters::whitelist::TrackerWhitelistAdapter;
+use torrust_tracker_rest_api_runtime_adapter::v1::container::TrackerHttpApiCoreContainer;
 
 use super::context::{auth_key, stats, torrent, whitelist};
