+++ packages/axum-rest-api-server/tests/server/v1/contract/context/torrent.rs
@@ -6,7 +6,7 @@ use torrust_tracker_primitives::peer::fixture::PeerBuilder;
 use torrust_tracker_rest_api_client::common::http::{Query, QueryParam};
 use torrust_tracker_rest_api_client::v1::client::{Client, headers_with_request_id};
 use torrust_tracker_rest_api_protocol::v1::context::torrent::resources::torrent::{self, Torrent};
-use torrust_tracker_rest_api_runtime_adapter::conversion;
+use torrust_tracker_rest_api_runtime_adapter::v1::conversion;
 use torrust_tracker_test_helpers::logging::logs_contains_a_line_with;
 use torrust_tracker_test_helpers::{configuration, logging};
 use uuid::Uuid;
