+++ packages/rest-api-application/src/v1/mod.rs
@@ -1,0 +1,5 @@
+//! Version 1 of the Torrust Tracker REST API application layer.
+//!
+//! This module contains all v1-specific port traits and use-case services.
+pub mod ports;
+pub mod use_cases;
