+++ packages/rest-api-application/src/v1/use_cases/torrent.rs
@@ -3,7 +3,7 @@ use torrust_info_hash::InfoHash;
 use torrust_tracker_primitives::pagination::Pagination;
 use torrust_tracker_rest_api_protocol::v1::context::torrent::resources::torrent::{ListItem, Torrent};
 
-use crate::ports::torrent::TorrentQueryPort;
+use crate::v1::ports::torrent::TorrentQueryPort;
 
 /// Use-case service for torrent-related API operations.
 ///
