+++ packages/rest-api-runtime-adapter/src/v1/adapters/torrent.rs
@@ -6,7 +6,7 @@ use torrust_info_hash::InfoHash;
 use torrust_tracker_core::torrent::repository::in_memory::InMemoryTorrentRepository;
 use torrust_tracker_core::torrent::services;
 use torrust_tracker_primitives::pagination::Pagination;
-use torrust_tracker_rest_api_application::ports::torrent::TorrentQueryPort;
+use torrust_tracker_rest_api_application::v1::ports::torrent::TorrentQueryPort;
 use torrust_tracker_rest_api_protocol::v1::context::torrent::resources::torrent::{ListItem, Torrent};
 
 use super::super::conversion;
