+++ packages/rest-api-runtime-adapter/src/v1/mod.rs
@@ -1,0 +1,7 @@
+//! Version 1 of the Torrust Tracker REST API runtime adapter.
+//!
+//! This module contains all v1-specific adapter implementations,
+//! the dependency injection container, and domain→protocol DTO conversions.
+pub mod adapters;
+pub mod container;
+pub mod conversion;
